Add original and "deroffed" versions of a couple of man pages
Included `find (1)` because it's a large, well-documented command and so I
assume the man page will be equally thorough, and a good test case; and the
sole manpage I could find in the `fr.ISO8859-1` section as a test of multi-
lingual capabilities.

(Since the source code of `deroff.py` includes some assumptions that input
text is `latin1`-encoded, I thought that was safest)

These were both downloaded from http://manpages.ubuntu.com/manpages/cosmic
and run through the version of `deroff.py` on `fish-shell` GitHub.