Implemented `Deroff::condputs`
I combined `condputs` and `condputs_tr`, because one overloads the other
later on. I also changed the type of `Deroff:self.tr` from `String` to
`Option<HashMap<u8, char>>` in order to be able to interface with
`util:translate` and `util:maketrans`. In addition, I fixed the type of
`Deroff:self.output` to a `String`