ci: stop running nightly in ci for now
Wouldn't be _bad_ per se to be running just the tests on nightly, but
the formatting job causes too many spurious errors (and a quick search
didn't provide me with the Travis-foo to exclude the `cargo fmt` job for
nightly).

Should look at re-adding post-MVP.

Also removes GitHub action for now. Need to keep things easy and simple,
but unfortunately the Rust version in the built-in Rust workflow is
out-dated.