feat: impl `condputs` (and `condputs_tr` (#64)
* Added `TranslationTable::new` and `TranslationTable::translate`

These functions mirror their python counterparts.

`TranslationTable::new` takes two equal length Strings and creates a
table for use in the translate function.

`TranslationTable::translate` takes a String and remaps the String based
on the table.

* Implemented `Deroff::condputs`

I combined `condputs` and `condputs_tr`, because one overloads
the other later on. I also changed the type of `Deroff:self.tr`
from `String` to `Option<TranslationTable>` in order to be able to
interface with `TranslationTable`. In addition, I fixed the type of
`Deroff:self.output` to a `String`

Formerly, the condputs table check would have no effect on output, so I
fixed that, as well as adding tests for `Deroffer::condputs`

Co-Authored-By: Ivan Tham <pickfire@riseup.net>

* have translate take &str instead of String

Co-Authored-By: Scott Steele <scottlsteele@gmail.com>

* Update src/deroff.rs

Co-Authored-By: Scott Steele <scottlsteele@gmail.com>

* fmt + fix tests to make translations take &str

* fixup! remove duplicate definition

Co-authored-by: Scott Steele <scottlsteele@gmail.com>
Co-authored-by: Ivan Tham <pickfire@riseup.net>