fix(drain): panic when different length
Python slice behavior works even when the end length is not achieved
but just the output length will be shorter but rust length requires it
to hit exactly that length and when the length is not enough or within
a boundary, it will panic. So we have to count by character boundary
and at the same time prevent the characters length from overflowing.